Get vulkan cts compiling
It's a bit hacky but it works. It needs to be run in a jail or
some system that does not have glslang and friends installed.

Something is adding a -I/usr/local/include flag when building the
external/* sources, which causes them to fail to build. By running
in a jail (or clean system without the packages corresponding to
external/* sources) there is nothing in /usr/local/include to
interfere with the build